requirements_streamlit_cloud.txt
streamlit
numpy
Pillow
tensorflow
fpdf